+++ nodes/curve/intersect_surface_plane.py
@@ -10,9 +10,9 @@ from sverchok.data_structure import updateNode, zip_long_repeat, match_long_repe
 from sverchok.utils.logging import info, exception
 from sverchok.utils.surface import SvSurface
 from sverchok.utils.geom import PlaneEquation
+from sverchok.dependencies import scipy, skimage
 
-from sverchok_extra.dependencies import scipy, skimage
-from sverchok_extra.utils.geom import intersect_surface_plane_msquares, intersect_surface_plane_uv
+from sverchok_extra.utils.manifolds import intersect_surface_plane_msquares, intersect_surface_plane_uv
 
 if skimage is not None or scipy is not None:
