+++ nodes/curve/marching_squares.py
@@ -8,9 +8,8 @@ from mathutils import Vector, Matrix
 from sverchok.node_tree import SverchCustomTreeNode, throttled
 from sverchok.data_structure import updateNode, zip_long_repeat, fullList, match_long_repeat
 from sverchok.utils.logging import info, exception
-
-from sverchok_extra.dependencies import skimage
-from sverchok_extra.utils.marching_squares import make_contours
+from sverchok.dependencies import skimage
+from sverchok.utils.marching_squares import make_contours
 
 if skimage is not None:
     from skimage import measure
