+++ nodes/curve/marching_squares_on_surface.py
@@ -10,9 +10,8 @@ from sverchok.data_structure import updateNode, zip_long_repeat, ensure_nesting_
 from sverchok.utils.logging import info, exception
 from sverchok.utils.field.scalar import SvScalarField
 from sverchok.utils.surface import SvSurface
-
-from sverchok_extra.dependencies import skimage
-from sverchok_extra.utils.marching_squares import make_contours
+from sverchok.utils.marching_squares import make_contours
+from sverchok.dependencies import skimage
 
 if skimage is not None:
     from skimage import measure
