+++ nodes/curve/ortho_project.py
@@ -9,9 +9,8 @@ from sverchok.node_tree import SverchCustomTreeNode, throttled
 from sverchok.data_structure import updateNode, zip_long_repeat, ensure_nesting_level, get_data_nesting_level
 from sverchok.utils.logging import info, exception
 from sverchok.utils.curve import SvCurve
-
-from sverchok_extra.dependencies import scipy
-from sverchok_extra.utils.geom import ortho_project_curve
+from sverchok.dependencies import scipy
+from sverchok.utils.manifolds import ortho_project_curve
 
 if scipy is not None:
