+++ nodes/surface/ortho_project.py
@@ -10,8 +10,7 @@ from sverchok.data_structure import updateNode, zip_long_repeat, ensure_nesting_
 from sverchok.utils.logging import info, exception
 from sverchok.utils.surface import SvSurface
 from sverchok.dependencies import scipy
-
-from sverchok_extra.utils.geom import ortho_project_surface
+from sverchok.utils.manifolds import ortho_project_surface
 
 if scipy is not None:
